// src/shared/api/generated/models/projectMembershipListParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectMembershipListStatus } from "./projectMembershipListStatus";

export type ProjectMembershipListParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * @minLength 1
   * @maxLength 512
   */
  cursor?: string;
  status?: ProjectMembershipListStatus;
};